requirements.txt
# Gaming AI Assistant Requirements

# Core dependencies
psutil>=5.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI Integration
openai>=1.3.0
anthropic>=0.7.0

# GUI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Additional utilities
python-dotenv>=1.0.0
pywin32>=306; platform_system=="Windows"
